+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v0.8.1
 
-* nothing yet
+## Fixed Issues
+
+* Fixed an issue where different formatted parameters having the same value in all cases were collapsed to a single parameter [#104](https://github.com/TNG/JGiven/issues/104)
 
 # v0.8.0